Reset eBMC reboot counter on graceful reboot
There needs to be a way to reset the reboot counter on graceful reboot
since migrating from IPMI. There is a PLDM PDR for the graceful reboot
counter that this commit will utilize to maintain functionality that
IPMI used to cover.

The call to reset the counter is disabled in this commit and will
be enabled in a follow-up commit once everything on the BMC side is
fully operational.

Change-Id: Ib89bc8826b5453b46a135868c455bdf1fb7b7c96
RTC:293028
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124997
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>