Changes to Gard record logic to keep duplicate deconfig records
This commit adds support for eBMC resource recovery.
Currently, eBMC sends all GARD Records to be applied to SBE
on reconfig loops. This is not desirable since if Hostboot
performs resource recovery on garded targets and sets their
HWAS state to functional, eBMC will still blindly apply all
gard records on a reconfig, and SBE will still see the targets
as non-functional. To remedy this behavior, an eBMC change will
be made to only apply GARD_Reconfig gards on reconfig loops.
To support this change, Hostboot needs to create GARD_Reconfig
gards for all gards we want to eBMC to apply.

Changes for gard record logic:
Before, only the first gard record for a target was kept.
Now we will keep up to 2 per target, one of each of the following:
  - GARD_Reconfig, to be used as Deconfig records for eBMC
    to apply on reconfig loops for resource recovery
  - one of any other GARD_ErrorType to be used by Hostboot as
    a normal gard record

Also, when applying Gard records for targets in host_gard, a
Deconfig record for the target will be created for it if there
is not one already. This is so that any gards Hostboot applies
will still be recognised by eBMC on reconfig loops.

Change-Id: Ie4345280ff78568eeeeec0f8feb56e57c3a42f1a
RTC: 281864
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124693
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>