Simplify Checks for Handling SBE Fails in Hostboot
This commit removes a check to see if the SBE had booted before calling
the sbe_run_extract_rc HWP.  Previously, the hostboot code falsely believed
that calling the HWP when the SBE thought it had been booted would cause
issues.  Now Hostboot will call the HWP in most scenarios to extract the
relevant fail informatioon.

This commit also traces and returns which Measurement Seeprom a
processor booted from in the same function that traces and returns
which Boot Seeprom the processor booted from.

Change-Id: I63c4b9af20607e537ba54b4c465748a719e0f769
CQ:SW537939
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124722
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>