PRD: LPC timeout guard issue with PNOR callout
The HWSV interfaces used to callout the PNOR and DPSS were called during
analysis before the error log was properly configured in the RasServices
code. Therefore, the log used in the callout interfaces were marked as
informational instead of the expected unrecoverable at the end of
analysis. The solution is to wait until RasServices to apply the
callouts.

Change-Id: I7684470860b6df85e38db39e5f50284fb24d9640
CQ: SW538803
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124891
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124955
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>