Remove support for writing VERSION partition data to errorlogs
Prior to this commit hostboot would lookup the VERSION pnor partition
and write its contents as a user details section in all error logs.
The BMC should be writing the single image data to PELs so this should
no longer be required. Removing this redundent behavior allows Hostboot
to avoid storing the the VERSION data in reserved memory as a fake lid.

Change-Id: Ib339275831283fd316133d40ee73b6a8fcf0a5a3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124948
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>