Add dynamic control of MAGIC_BREAK
It is often useful to add a MAGIC_BREAK into the code to catch
certain paths of code.  However, some of those paths can run so
often to make it nearly useless to break on every execution.

This change adds a new simics variable "hb_ignorebreak" that
controls the behavior of MAGIC_BREAK.  When hb_ignorebreak==1,
the MAGIC_BREAK will be ignored.  The default behavior will be
to honor MAGIC_BREAK (hb_ignorebreak==0).

Change-Id: I381ccf3cd8a5dadeb693fa5b8e85025681b37613
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125066
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>