Do not remove PNOR Gard actions for altpnor failures
A previous set of logic was removing gard and deconfig actions
from the logs that are returned from alternate pnor accesses.
The point of this was to avoid losing a processor due to any
of these failures (which was possible based on the way errors
can be presented).  Later on we added a gard of the PNOR part
but we didn't remove the ungard logic.

The change here is to explicitly remove the service actions
for the processor but leave anything else (specifically the
PNOR) intact.

Change-Id: I22d9a15fe0b160187d3978dfb7bdde6d9c04ce9c
CQ: SW539608
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125147
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>