Resolve bus callouts to intermediate FRUs (MRW processing, commit 1/2)
When an error occurs over a hardware bus (such as an XBUS or SPI
connection), any field-replaceable unit (FRU) between the two
endpoints of the bus connection could be the cause of the
problem. Therefore, when a Hostboot error calls out bus endpoints, the
system should also call out each FRU that is in between the endpoints.

On FSP systems, Hardware Server adds these intermediate callouts. For
BMC systems, Hostboot needs to add them. This commit is the first of
two that implement that functionality.

The intermediate FRUs are provided in the FRU_PATH tag in the MRW bus
connection descriptions. See BusFruCallouts.pm for an example.

Change-Id: Ieef74bbf18e54744297935b2f991bf6f2e07aea9
RTC: 122928
CMVC-Prereq: 1145088
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125248
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>