PLDM: Implementation of the getPLDMVersion request handler
The getPLDMVersion request is part of the Discovery Control message.
Added a Discovery Control Message to the PLDM messages to be handled
and implemented the getPLDMVersion request handler.

The handler also replies with 'unsupported' because only using message
as an indication that the PLDM daemon has been restarted, not interested
in the message itself.

Change-Id: I898fce6c337a56b1eaf6205afd7b3fde523f8198
RTC:288391
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125295
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>