Tweak the SOL console output for fake gard records
Don't emit a console message if the GARD is GARD_Reconfig since that
type is considered a "fake" GARD record. Emitting a message as if a GARD
is being applied in that case leads to confusion since no GARD is
actually being created. Its actually just used as a means to remember to
deconfig.

Change-Id: I7dfae9f2e4b820dfa3f632507a52b57ff14cffa5
CQ:SW539341
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125326
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>