Consolidate PLDM response validation logic
Prior to this change all of the PLDM request APIs had if-checks
in the function that would generate an error if data in the
response was not we expected. This commit moves that logic into
a function which each PLDM request API can call. This reduces
redundency and will help in the future if something needs
to change because it can be done in one place.

Change-Id: Ie3565120018bec5ecfefe92d73536fff453058d8
RTC: 248498
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125330
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>