Fix istep error logs so that the IPL stops
When an istep encounters errors they are aggregated into a common
IstepError object.  That object is then used to return an error
log to the istepdispatcher logic.  The interface to retrieve the
stored log is destructive in that it will clear the internal
pointer and give it to the caller.  Therefore if the caller does
not consume the result the istep aggregation log will be lost.

In these problematic cases multiple getErrorHandle() calls were
made, which leads to a nullptr being returned to istepdispatcher,
which allows us to continue the IPL when we should not.

Change-Id: I78990bccf418e470089d7b51b42710caad55c086
CQ: SW539808
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125365
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>