PLDM: Doing a shutdown when receiving a getPLDMVersion request
Doing a shutdown of HB when receiving a getPLDVersion request after
HB has started the PDR exchange with the PLDM.  Receiving a getPLDMVersion
request is indicative of the PLDM daemon being restarted/reloaded which
will reset the HB PLDM sequencing *after* the PDR exchange has started
between the PLDM and HB. Resetting the sequencing will cause
issues for HB which are hard to recover from, so instead of attempting
recovery, HB will shut down.

Added new shutdown status, SHUTDOWN_STATUS_PLDM_RESET_DETECTED' that
will be passed as reason for the shutdown.

Added attribute HALT_ON_BMC_PLDM_RESET that will be set to
1, true, once a critical exchange with the PLDM commences, such as
the exchange of the PDRs. This will be used as a flag, for when a
getPLDMVersion is received, to determine if a shutdown is needed
or the message can be ignored.

Change-Id: I3bf6389083cb783aac28f6ae7fe4c98eba6a7ac7
RTC:288391
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125425
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>