Change compile flag for building PLDM_DEPS in the pnor makefile
Previously we used the CONFIG_FILE_XFER_VIA_PLDM flag to decide
whether or not to compile the PLDM_DEPS in the PNOR module. This
causes some linking problems if CONFIG_PLDM is set but
CONFIG_FILE_XFER_VIA_PLDM is not. It will be safe to compile
this CONFIG_PLDM is set so we will use that flag instead.

Change-Id: I3173263acad5eb6d21bc1b46b0434a01910e5af0
RTC: 292695
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125501
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>