Enable Resource Recovery for garded TPM targets
Currently, If an eBMC system has hb_tpm_required set to 'Required',
but the primary TPM is manually garded out, the system will not
recover the guarded out TPM when 'obmcutil recoveryon' is set.
This commit sets ATTR_BLOCK_SPEC_DECONFIG=1 in the
RC_TPM_NOFUNCTIONALTPM_FAIL path so that recoverable gards for TPM
will not be applied on the following reboot.

Change-Id: I9ca764e1c4f1bec06cbd87ff52edc6a40e8b8728
CQ:SW540175
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125815
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>