Fix resource deallocation bug in the trace weave tool
Previously, there was a bug in the weave tool in the cleanup_strings
resource deallocation routine. It was iterating from 0 to the size of
an std::map and deallocating the strings contained in it, using the loop
counter as the index. However, the map was sparsely populated; the
keys were Hostboot trace hashes, so the map was unlikely to contain
the keys 0, 1, 2... etc.

The act of accessing the map at index [i] was inserting an element
into the map, thereby adding 1 to the size of the map, causing the
loop to essentially create elements in the table with keys 0 up to
whatever the largest hostboot trace hash key was, and the act of
approaching the loop's termination condition caused it to move one
further away. This eventually consumes all the memory on the system
and crashes the process or the machine.

This commit fixes the issue by not cleaning up the strings manually
and letting the OS reclaim the memory itself.

Change-Id: I81f1e151cb8c445058d4dfb60a1429eae4ea3650
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125615
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>