TI with RC if an error occurs during pldm_base_init
Prior to this change Hostboot would return an error from
pldm_base_init. If Hostboot populates the Errl parameter from
a function passed to TASK_ENTRY_MACRO Hostboot will shutdown with
an error log. Since PLDM failed to init, the error log will not
make it to the BMC so it will be lost. Instead Hostboot wants the BMC
to dump the host. Hostboot can cause this to happen by TI'ing with
an RC. This commit will force Hostboot to TI with an RC if an error
is encountered in PLDM init.

Change-Id: I9dd0200f10f8437d41dd55a2f924144225e6d9d4
RTC: 248498
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125640
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>