eBMC - Real Time/Date Support: ErrlManager Changes
This commit introduces changes to ErrlManager related to the new
way of fetching real date/time from BMC via PLDM. The base date/time
is fetched from BMC once during the start of the PLDM daemon, and a
timebase (number of ticks) is noted at that time. Whenever a current
date/time is needed, the new timebase is taken, the differece between
the current and base timebase is calculated and converted to the number
of seconds elapsed, and the seconds are added to the base date/time.

Change-Id: I31fbbbc96d23dc24c34a7675991df1690e0b84b8
RTC: 255972
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125874
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>