VPD: Verify record is valid after ECC data validation completed
The ECC data validation code only validates that the ECC data to the
record data as a blob.  It is does not verify that the record itself is
valid.  This became an issue when forcing the ECC update to update
the ECC data when it was incorrect for a given record and the record
data was garbage.  Garbage in - garbage out but from the ECC validation
code this was correct.

The verify record methods will verify that the record is not complete garbage
by comparing the record's meta data to what is found in the record
itself.  If the record itself has corresponding record name and
size as found in the meta data, then the record is considered verified.
If not then the record is not verified and will throw an error which
can be used to get the next MVPD data.

Change-Id: Ica456f41985d7409700fe83bb0caab84f0d7cf5b
CQ: SW540633
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125900
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>