Add procedure callout for node comm failures
During the node-to-node communication between Hostboot instances
it is possible that the other side of the connection could have
checkstopped.  There is a moment in step18 where the nodes are
physically connected and able to carry some traffic, but the
IOVALID is not set so they won't carry the checkstop signal.

This change simply adds a procedure callout that will direct
the servicer to look for other logs from the FSP that indicate
a checkstop condition.

Change-Id: I5ee3eb1d25fa29a5f1a674d752d7ead31421e91f
CQ: SW537455
CMVC-Prereq: 1145397
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126023
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>