SBE CUSTOM IMAGE: Removed any reference and calls to method selectBestCores
The method selectBestCores was called to select a core for a SBE custom
image with limited space requirements.  Those space requirements no
longer exist for P10 therefore it is no longer necessary to select the
"best" core.  There should be plenty of space allocated for any and all
cores.

Having said that, removing the call to selectBestCores has greatly simplified
the code. selectBestCores was setting a core mask that was used by the HWP
call p10_ipl_customize.  Now the call to p10_ipl_customize will use a core
mask that has all available cores the system has.  The code reduces to:
1) Confirm that the desired minimum cores does not exceed the max cores on
   system. If it does, create an error log stating so and return error to caller
2) Call the HWP call to create a customized SBE with any and all cores
   available.  If the HWP fails, then return error to caller
3) If the HWP call was successful, verify that it was able to create a custom
   SBE image, at a minimum, with the number of desired cores.  If not return an
   error to caller

Change-Id: I649c0c9a6ddab7075db0ec80de170966bdb21c64
CQ: SW540381
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126152
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>