Updated validation check on an SPD to determine if it is for an MDS DDIMM
After a conversation with Kevin McIlvain and Steven Janssen it turns
out the validation check on an SPD to determine if it is indeed for
an MDS DDIMM was incorrect.

In light of the new information on how to determine if working with
an MDS DDIMM:
* byte 2 = 0x0C
* byte 3 = 0xBA
* byte 264 = 0x15

This file has been updated to correctly check if the SPD is for
and MDS DDIMM or not.

Change-Id: Ia668e0b8f1e630e4800c8ea2b0e68edada6a8f2c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126241
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126242
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>