SPD: Correcting the record
After a conversation with Kevin McIlvain, Steven Jansen and looking over
the SPD spec, there is no *memory* type values 0x3A(MDS) and 0x0A(DDIMM).
There is however a *module* type 0x0A for DDDIM and type 0xBA for a Type
1 DDIMM which is being used to determine if the SPD is an MDS DDIMM or
not.

Therefore dropping the definitions for memory type 0x3A and 0xA and adding
module type 0xBA for the MDS.

The memory type is defined in byte 2 and the module type is defined in
byte 3 of the SPD.

Also fixed the fallout from making these changes which means updating
files to use the correct enumerations.

Change-Id: I0f621b7628393fc8503264e7027f325a529f22b6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126243
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>