Refactor ATTR_DEFAULT_HB_CORE -> ATTR_HB_FALLBACK_CORES (1/1 HB)
The original design of ATTR_DEFAULT_HB_CORE was not comprehensive
enough to cover fused vs not-fused cores as well as valid backing
cache. To convey this additional info ATTR_DEFAULT_HB_CORE will be
renamed to ATTR_HB_FALLBACK_CORES and will instead provide a bit
mask of the cores HB determines are valid.

This is the commit on HB's side to customise the value for
ATTR_HB_FALLBACK_CORES.

Change-Id: If39f2f72cca4ef1fb7fdb96465d7d380d3d987cf
CMVC-Prereq: 1146569
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126306
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>