Compute ATTR_HOSTSVC_PLID During IPL
This commit intoduces changes to compute the value of ATTR_HOSTSVC_PLID
based on the highest runtime PLID number found in the HBEL partition
during the initialization of PNOR at IPL-time. The attribute is then
passed on to runtime and runtime uses it as the base for the current
PLID number. That ensures that HBRT PLIDs don't repeat on system
reboots and that unique PLIDs are sent to BMC.

Additionally, this commit has changes to allow for PNOR writes to HBEL
partition during runtime.

Change-Id: I4bc79e944659d540f72440775de9a2d3603a0aad
RTC: 294833
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126716
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>