Set the attribute FW_RELEASE_VERSION at runtime
Setting the FW_RELEASE_VERSION at runtime in case there
is an update to the code that changes the FW Release Version.
This will keep the version info in sync with the updated code.

Moved the Marker LID structure from the file utillidmgr.H to the
errlmanager.H to be inline with where it is being used.

Added function setFwReleaseVersion_rt that does the setting of
the FW Release Version after retrieving it from the Marker LID.

Change-Id: I1f6fbcc4b11098072bdc0065ce1e84ee774e1303
RTC: 282070
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126603
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>