Runtime dynamic deallocation of guarded resources (commit 1/2)
Previously, when PRD, running within HBRT, detected an error in a
piece of hardware, HBRT would create a guard record (via the error log
callout) for that hardware and send it to the BMC. However, the
hypervisor would not be notified of the error and would keep using the
hardware like before.

This commit adds functionality to notify the hypervisor when a piece
of hardware is guarded out, so that it can stop using that hardware if
possible to avoid future errors.

There will be a followup to this commit to add testcases.

Change-Id: Ic23f84420fc55a86fdb5d509c693dee382b56590
RTC: 296408
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126687
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>