Push EECACHE updates into PNOR at runtime
Currently we exclusively use a reserved memory area to read and
write the EECACHE data at runtime.  This is a problem because we
lose track of any VPD writes that happen.  Note that this is only
a problem on BMC systems because we rely on the FSP to persist
a flag that tells us to refresh the data on the next reboot.

The fix here involves a couple tweaks:
1) Instead of preloading a reserved memory section, we will pull
the data from PNOR directly as HBRT loads.
2) Add a call to PNOR::flush() after every write.

Change-Id: Id3408aed912f4e3354e51d388ea7c675d80019de
CQ: SW542320
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126755
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>