setup script for eBMC standalone patching
Adds a new script which can setup the hostboot standalone repo for
building eBMC patches given a supplied package name. It has other
options that allow for it to also replace the simics xml with the system
xml for targeting patching and it can also just copy the system xmls
from the pnor build from the package for inspection.

Change-Id: I539e684b646e441a0b04efc281e2bcc34871be0e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/126999
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Jayashankar Padath <jayashankar.padath@in.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>