Support for precompiled VPD ECC implementations
This commit adds support for injecting precompiled ECC implementations
into the Hostboot build process.

If a user puts a file named libecc_static.a in the root directory of a
Hostboot build, that file will be used as the implementation of
Hostboot's VPD ECC functions. Otherwise, Hostboot will fetch and
compile the default ECC implementation from source.

RTC: 295900
Change-Id: I4b98637b55d38581a3f319c33716ee4f21af6a98
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127173
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>