Runtime interface to provide BPM lifetime stats for NVDIMMs
This will implement a runtime (HBRT) interface that can be called
to provide the lifetime percentage value from the BPMs.  HBRT
will be invoked via firmware_notify, then read the data from
the hardware register on the NVDIMM, and finally send the
results back to the host via a firmware_request.

Change-Id: I6a53620ce003c84fc90560f3cfb319847d88decd
RTC: 297022
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127108
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>